requirements.txt
streamlit
groq
python-dotenv
sentence-transformers
faiss-cpu
torch
numpy
pytesseract
python-docx
PyMuPDF
pdf2image
Pillow
opencv-python-headless